% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-package.R
\docType{package}
\name{profile-package}
\alias{profile}
\alias{profile-package}
\title{profile: Read, Manipulate, and Write Profiler Data}
\description{
The profiler package defines a stable data format for profiler data,
see \code{\link[=validate_profile]{validate_profile()}} for a definition.
It supports reading and writing \link{Rprof} \code{.out} and \href{https://github.com/google/pprof}{pprof} \code{.proto} files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-prof.github.io/profile/}
  \item \url{https://github.com/r-prof/profile}
  \item Report bugs at \url{https://github.com/r-prof/profile/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
