% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler-package.R
\docType{package}
\name{profiler-package}
\alias{profiler}
\alias{profiler-package}
\title{profiler: Read, Manipulate, and Write Profiler Data}
\description{
Defines a data structure for profiler data, and methods to read and
write from the 'Rprof' and 'pprof' proto file formats.
}
\examples{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/profiler}
  \item \url{http://krlmlr.github.io/profiler}
  \item Report bugs at \url{https://github.com/krlmlr/profiler/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
