% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof-write.R
\name{write_pprof}
\alias{write_pprof}
\title{Write profiler data to a proto file}
\usage{
write_pprof(path, ds)
}
\arguments{
\item{path}{A file name}

\item{ds}{Profiler data}
}
\description{
Write profiler data to a proto file
}
