% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprof-write.R
\name{write_rprof}
\alias{write_rprof}
\title{Write profiler data to an R profiler file}
\usage{
write_rprof(path, ds)
}
\arguments{
\item{path}{A file name}

\item{ds}{Profiler data}
}
\description{
Write profiler data to an R profiler file
}
