% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-package.R
\docType{package}
\name{profile-package}
\alias{profile}
\alias{profile-package}
\title{The profiler package defines a stable data format for profiler data,
see \code{\link[=validate_profile]{validate_profile()}} for a definition.
It supports reading and writing \link{Rprof} \code{.out} and
\href{https://github.com/google/pprof}{pprof} \code{.proto} files.}
\description{
The profiler package defines a stable data format for profiler data,
see \code{\link[=validate_profile]{validate_profile()}} for a definition.
It supports reading and writing \link{Rprof} \code{.out} and
\href{https://github.com/google/pprof}{pprof} \code{.proto} files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-prof/profile}
  \item \url{http://r-prof.github.io/profile}
  \item Report bugs at \url{https://github.com/r-prof/profile/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
