% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-package.R
\docType{package}
\name{profile-package}
\alias{profile}
\alias{profile-package}
\title{profile: Read, Manipulate, and Write Profiler Data}
\description{
Defines a data structure for profiler data, and methods to read and
write from the 'Rprof' and 'pprof' proto file formats.
}
\examples{

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/profile}
  \item \url{http://krlmlr.github.io/profile}
  \item Report bugs at \url{https://github.com/krlmlr/profile/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
