% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprof-read.R
\name{read_rprof}
\alias{read_rprof}
\title{Read profiler data from an R profiler file}
\usage{
read_rprof(path)
}
\arguments{
\item{path}{A file name}
}
\description{
Read profiler data from an R profiler file
}
