% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprof-read.R
\name{read_rprof}
\alias{read_rprof}
\title{Read profiler data from an R profiler file}
\usage{
read_rprof(path, ..., version = "1.0")
}
\arguments{
\item{path}{A file name}

\item{...}{Ignored}

\item{version}{Version of the data, currently only \code{"1.0"} is supported.
Pass an explicit value to this argument if your code depends on the data
format.}
}
\description{
Read profiler data from an R profiler file
}
