% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof-write.R
\name{write_pprof}
\alias{write_pprof}
\title{Write profiler data to a proto file}
\usage{
write_pprof(ds, path)
}
\arguments{
\item{ds}{Profiler data, see \code{\link[=validate_profile]{validate_profile()}}}

\item{path}{Target file name}
}
\description{
Use the \href{https://github.com/google/pprof}{pprof program} to read and
analyze the file created by this function.
}
