% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprof-write.R
\name{write_rprof}
\alias{write_rprof}
\title{Write profiler data to an R profiler file}
\usage{
write_rprof(ds, path)
}
\arguments{
\item{ds}{Profiler data, see \code{\link[=validate_profile]{validate_profile()}}}

\item{path}{Target file name}
}
\description{
Use the profvis or proftools R packages to further analyze files created by
this function.
}
